mfa_project/requirements.txt
Flask==2.3.3
flask-jwt-extended==4.6.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
pymongo==4.6.1
pyotp==2.9.0
qrcode==7.4.2
face-recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
email-validator==2.1.0.post1
